piece move logic

+++ src/lib.rs
@@ -2,6 +2,6 @@ pub mod board;
 pub mod consts {
     pub const HEIGHT: i32 = 700;
     pub const WIDTH: i32 = 700;
-    pub const ROWS: u8 = 8;
-    pub const COLS: u8 = 8;
+    pub const ROWS: usize = 8;
+    pub const COLS: usize = 8;
 }
